pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "pith"
version = "2.0.0"
description = "Visual markdown hierarchy editor with prose analysis CLI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pyyaml>=6.0.2",
    "markdown>=3.7",
    "python-frontmatter>=1.1.0",
    "python-multipart>=0.0.9",
    "aiofiles>=24.1.0",
    "pywebview>=5.0",
]

[project.scripts]
pith = "backend.launcher:main"

[tool.setuptools.packages.find]
include = ["backend*"]
